Remove the machinery that existed because there was no crontab
A real crontab is now a requirement, not a hope: public/cron.php writes a
heartbeat, Core\Scheduler\CronHealth reports it from three independent
traces, and SetupController::save() refuses a first-run install while it is
not verifiably active. Everything this project built to advance its queue
WITHOUT one is therefore redundant complexity on the path — unattended
self-update — that has caused the most production outages.

Deleted: Scheduler\SchedulerContinuation, SchedulerContinuationRoute,
SchedulerKick, SliceOutcome, Database\MigrationChain, Http\SelfRequest
(its only two callers were the two chains), the
POST /api/scheduler/continue route and its controller, and the poor man's
cron at the foot of public/index.php.

Nothing was lost with that tail: public/cron.php already runs the scheduler
pass, the journal cleanup, LoginThrottler::purgeStale() and
PdfThumbnailCache::purgeStale() on every pass, which is exactly what the
tail did once a minute at a visitor's expense.

MaintenanceGate keeps gating every path — its exemption for the
continuation route is dead with the route, so checkBlocking() no longer
takes a path and FrontController no longer threads one. The pending
migration short-circuit, the progress page, POST /api/system/migration-step
and its touchInProgress() heartbeat all stay: a visitor arriving mid-
migration still needs them. Only the self-hops are gone.
InstallUpdateHandler and RestoreBackupHandler still refuse to migrate in
the process that replaced the files; they simply let the next cron pass
pick the resume task up instead of kicking one.

Core settings are never pruned — SettingService::pruneUndeclared() is
scoped to a module_id and nothing touches a `module_id IS NULL` row that
stopped being registered — so dropping the register() calls alone would
have left seven orphan rows on every installed site. Hence
SettingRepository::deleteCoreSettings() and a one-time guarded cleanup in
public/index.php, in the same shape as the notifications-v2 block.

The end-to-end harness no longer provisions scheduler_max_hops at zero; it
drives the queue purely through its public/cron.php subprocess, and the two
scenarios that used to wait the poor man's cron out now call runScheduler()
explicitly.
