+++ core/Http/FrontController.php
@@ -79,10 +79,7 @@ class FrontController
         // future non-web entry point) — never blocking in that case.
         if ($this->maintenanceGate !== null) {
             $bypassRequested = $request->getQuery(MaintenanceGate::BYPASS_QUERY_PARAM) !== null;
-            // The path goes in because one route must never be gated: the
-            // scheduler's continuation endpoint, which is how an update in
-            // 'migrating' gets finished at all. See MaintenanceGate.
-            $blockingUpdate = $this->maintenanceGate->checkBlocking($bypassRequested, $request->getPath());
+            $blockingUpdate = $this->maintenanceGate->checkBlocking($bypassRequested);
             if ($blockingUpdate !== null) {
                 return $this->renderMaintenanceInProgress($blockingUpdate);
             }
