+++ core/Maintenance/MaintenanceGate.php
@@ -8,7 +8,6 @@ declare(strict_types=1);
 
 namespace Core\Maintenance;
 
-use Core\Scheduler\SchedulerContinuationRoute;
 use Core\Security\AuthSession;
 use Core\Security\Role;
 
@@ -30,36 +29,18 @@ use Core\Security\Role;
  * bypass lets through, so a visitor who stumbles onto the parameter learns
  * nothing they couldn't otherwise reach once the update finishes.
  *
- * **One path is never gated: the scheduler's own continuation endpoint.**
- * That exemption is not a convenience, it is what lets an update finish at
- * all. Task\InstallUpdateHandler deliberately does not migrate in the
- * process that replaced the files: it sets the status to `migrating`,
- * queues the resume task, and calls Scheduler\SchedulerKick::now() so
- * another process picks it up immediately. That hop lands on
- * `POST /api/scheduler/continue` — and this gate, which runs before
- * routing and covered every path without exception, answered it with the
- * "update in progress" page, because the row it gates on IS the update the
- * hop exists to advance. The hop then fell through to the poor man's cron
- * at the end of public/index.php, which is throttled to once a minute and
- * had just been stamped by the pass that ran the install. So nothing ran,
- * and on an installation with no crontab the resume task waited for
- * whatever unrelated request happened to arrive next.
- *
- * Production, scoutmagic.be, 2026-08-31: `dev-afd995a → dev-a29f7ae`
- * queued its resume task at 15:03:00; it ran at 15:27:35 — and only
- * because UpdateHistoryRepository::findInProgress() had by then declared
- * the update abandoned all by itself, at exactly `started_at` plus fifteen
- * minutes, which is what ungated the site. The task then found a `failed`
- * row and gave up. Six updates died that way in forty-eight hours, each of
- * them serving this page to every visitor for the whole fifteen minutes,
- * and none of them was a slow migration: two of the last three changed no
- * schema at all. The wait was the gate, not the work.
- *
- * Exempting the route costs nothing. It is public, session-free and
- * authenticated by a shared secret checked inside
- * Http\Controller\SchedulerContinuationController, so a caller without the
- * secret gets a 403 rather than anything this gate was hiding — and the
- * gate is an availability gate, not a security boundary, as above.
+ * **Nothing is exempt from this gate any more, and that is a change.**
+ * One path used to be: `POST /api/scheduler/continue`, the hop the
+ * scheduler emitted onto its own site so an update stuck in `migrating`
+ * would be finished by another process. The gate answered that hop with
+ * the "update in progress" page — it gates on the very row the hop came
+ * to advance — and six updates died of the resulting wait in forty-eight
+ * hours on scoutmagic.be. The exemption fixed that; requiring a real
+ * crontab removed the need for it. `public/cron.php` runs outside the web
+ * request entirely, so no gate has ever been in its way, and the resume
+ * task is picked up by the next pass (at most a minute) whatever this
+ * gate answers a browser. The hop, its route and its exemption are all
+ * gone — see ARCHITECTURE.md § 8.5.
  */
 final class MaintenanceGate
 {
@@ -70,25 +51,12 @@ final class MaintenanceGate
     }
 
     /**
-     * @param string $path the requested path, so the one route that must
-     *        never be gated can be recognised — see the class docblock.
-     *        Defaults to the empty string for call sites that have no path
-     *        to offer (tests, any future non-web entry point); an empty
-     *        path is gated like any other.
-     *
      * @return UpdateHistory|null the in-progress update to show a
      *         maintenance page for, or null if the visitor should proceed
      *         to the normal site
      */
-    public function checkBlocking(bool $bypassRequested, string $path = ''): ?UpdateHistory
+    public function checkBlocking(bool $bypassRequested): ?UpdateHistory
     {
-        // Before findInProgress(), deliberately: that call also marks a
-        // stale row failed as a side effect, and a hop coming to finish an
-        // update has no business tripping the watchdog on it.
-        if ($path === SchedulerContinuationRoute::PATH) {
-            return null;
-        }
-
         $inProgress = $this->updateHistoryRepository->findInProgress();
         if ($inProgress === null) {
             return null;
